# Added Optimization and OverSampling