# Changes
Added OverSampling to increase the effective resolution from 10 bit up to a maximum of 15 bits based on the use of `int` (default: 12 bit)
- Practically the limit on the bit resolution is 13 bits which would take 50ms to complete the four oversampled analog reads.
- Any higher bit resolution would result in times longer than the input latency that @TomStanton measured during his testing of the controller (111ms).
  - 11 bit - 5ms
  - 12 bit - 13ms
  - 13 bit - 54ms
  - 14 bit - 221ms
  - 15 bit - 914ms